Added UserSession controller
Added a UserSession controller to store the users data thats held on
our servers. Added test variable on dashboard and added some well
thought out security sollutions to the login system and logging in with
a social account. Previously if someone was able to get ahold of even
one of a users social media accounts or even register someone for a
social account they may not have registered under there email. they
could use that account to login with Comrade and all of a sudden that
hacker would have access to ALL of there other social media accounts.
This is bad and we have developed 2 solid methods to prevent this from
happening. 1 of those methods has been implemented in this commit.