Added MORE security
We decided best that we don’t even hold the users accessToken since
everything is taken care of client side its best we don’t. Besides
there really useless after expired unless we used our own
implementation of oauth-shim to request new access tokens which is what
we can do with future funds from winning 59DaysOfCode. Actually one of
our Use Of Funds fro the winnings (if we i win) is to pay the creator
of HelloJS to help me further HelloJS integration into the app. I have
a list of things to do for this Framework that will help Comrade a lot.