Last Commit for 59DOC
sadly this is my last commit, was not able to get the app to a
presentable stage. By showcase i will have a production ready
application with quality design. However this is what the judges will
see, or not see, i have not decided. From here on out I’m creating a
new branch for me only.